Renamed rebuildunassigned to RebuildUnassigned, which is how it was called in the code. Added a new function to draw polygons and return their vertices, GetNewPolygon.m, which uses Matlab's impoly function. Fixed a few other case-sensitivity issues. One thing that I had to change, but not really sure why was how clusters are deleted and redrawn in ClusterFunctions.m:420 or so. Indexing into the cluster cell arrays seemed incorrect. Also fixed bug in how times were returned from LoadIndexSnip.m, they were the file position in bytes, not the time of the snippets in samples.
Verified that the code runs to completion, including loading, sorting
and removing cross-talk, removal of clusters from array plot, and
multi-dimensional clustering. I think there might still be something
strange going on with loadaibdata. The cross-talk that is shown looks
correct, but when it is appended and shown in the channel grouping
window, it doesn't look right. But not positive, and I don't know where
the bug is if it exists.

Either way, this seems to be all or almost all that we need for issue
9, milestone 1, getting the code running on new Matlab. Going to try to
debug the AIB data loading, and do some robustness, before moving on to
getting HDF I/O stuff working.